% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{HdDataverse}
\alias{HdDataverse}
\title{HdDataverse
This function takes in the File Persistent ID available for each data file in the Harvard Dataverse and user's API key.}
\usage{
HdDataverse(ID, APIkey)
}
\arguments{
\item{ID}{the File Persistent ID available in the Metadata section for each data file}

\item{APIkey}{users need to first obtain the free API authentication key from Harvard Dataverse}
}
\value{
a dataframe
}
\description{
HdDataverse
This function takes in the File Persistent ID available for each data file in the Harvard Dataverse and user's API key.
}
\examples{
# obtain the state_share data in the Startup Cartography Project with File Persistent ID
#' doi:10.7910/DVN/BMRPVH/0WQQMT
HarvardData <- HdDataverse("doi:10.7910/DVN/BMRPVH/0WQQMT", "HARVARDDATA_KEY")
# people could save the file in RData format
saveRDS(HarvardData, file = "Harvard_Dataverse_data.rds")
}
