% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legen.R
\docType{package}
\name{legen}
\alias{legen}
\title{legen: A package for collecting and studying the Drivers for Local Entrepreneurships}
\description{
The legen package provides four categories of important functions:
locsent, statepop, HdDataverse and patent.
}
\section{locsent functions}{

The foo functions ...
}

\section{statepop functions}{

The foo functions ...
}

\section{HdDataverse functions}{

The foo functions ...
}

\section{patent functions}{

The foo functions ...
}

