% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{locsent}
\alias{locsent}
\title{This function takes in google trend search keywords, beginning year, ending year as the keys and outputs annual time series data for each region within US.}
\usage{
locsent(key, begYr, endYr)
}
\arguments{
\item{key}{the input keywords that you would like to get the google trends for}

\item{begYr}{the year as the starting point for your time frame}

\item{endYr}{the year as the ending point for your time frame}
}
\value{
a dataframe
}
\description{
This function takes in google trend search keywords, beginning year, ending year as the keys and outputs annual time series data for each region within US.
}
\examples{
# obtain the google trends data for "corruption" from 2006 to 2016.
corruption <- locsent("corruption", 2006, 2016)
}
