% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{patent}
\alias{patent}
\title{patent
This function takes in the beginning year, ending year of the data and outputs the number of patent granted for each state from the time period specified.}
\usage{
patent(begYear = 1993, endYear = 2018)
}
\arguments{
\item{begYear}{the year as the starting point for your time frame}

\item{endYear}{the year as the ending point for your time frame}
}
\value{
a dataframe
}
\description{
patent
This function takes in the beginning year, ending year of the data and outputs the number of patent granted for each state from the time period specified.
}
\examples{
# obtain the number of patent granted for each state from 2004 to 2016
patentts <- patent(2004, 2016)
}
