% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{statepop}
\alias{statepop}
\title{statepop
This function takes in the user's API key, beginning year, ending year of the data and whether or not excluding puerto ricos.}
\usage{
statepop(APIkey, begYr = 2000, endYr = 2018, exclude_puerto_rico = TRUE)
}
\arguments{
\item{APIkey}{users need to first obtain the free API authentication key from US Census Bureau}

\item{begYr}{the year as the starting point for your time frame}

\item{endYr}{the year as the ending point for your time frame}

\item{exclude_puerto_rico}{return the population estimates without puerto rico population estimates (\code{TRUE}) or #' with (\code{FALSE})}
}
\value{
a dataframe
}
\description{
statepop
This function takes in the user's API key, beginning year, ending year of the data and whether or not excluding puerto ricos.
}
\examples{
# obtain the state population from 2014 to 2016 excluding puerto ricos
statepopulation <- statepop("CENSUS_KEY", 2014, 2016, TRUE)
# obtain the state population from 2000 to 2018 including puerto ricos
statepopulation <- statepop("CENSUS_KEY", FALSE)
}
